unify abstract_io.* with those from lilv.
When I was verifying the changes along with the approach I documented in
aap-lv2 README.md, it worked fine. When I pulled the binaries from the build,
it didn't. The reason was that I was making changes only in serd, not lilv,
and the changes were further copied into `lilv_direct` hack directory and
then were made changes. It was not reliable at all.

It is still not reliable here, but at least I am aware of the cause of the
problem now.